Fix crash on fopen error
Fixed an issue whereby the return value of fopen was not checked for a
NULL pointer causing a crash. This error occurred when opening a file
with the character "é", when it was the first file opened when starting
the program.
Error messages now report the errno string.
Added preliminary stream support.

Signed-off-by: Mahyar Koshkouei <deltabeard@users.noreply.github.com>